usb test application for usb btree class driver
This patch exports usb test application for usb btree class driver.
This application gets video stream from the connected btree board
and creates bmp file for the result.

This application only works when the class driver is loaded
and btree demo board is connected to the PC.

usage : ./usb_test /dev/btree%number
(the number of btree device that is connected to the PC)

Change-Id: Id5f71bb0fcd32910afc9ef0663ba2389f0157bba
Signed-off-by: HyejungKwon <cjscld15@nexell.co.kr>