v4l2-test application for Btree ISP
This patch exports v4l2-test application for btree isp.
This application gets video stream from the connected btree board
by usb and creates bmp file for the result.

This application only works when the class driver is loaded
and btree demo board is connected to the PC.

usage : sudo ./v4l2_test /dev/video%number
(the number of video device that is connected to the PC)

Change-Id: Ie040b771ace9bc5808f43800f24754355aff4226
Signed-off-by: HyejungKwon <cjscld15@nexell.co.kr>